completed 4.22

+++ part4/controllers/blogs.js
@@ -11,16 +11,9 @@ blogRouter.get('/', async (request, response) => {
   })
   
 blogRouter.post('/', async (request, response) => {
+
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-    if (!request.token || !decodedToken.id) {
-      return response
-        .status(401)
-        .json({ error: 'token missing or invalid' })
-    }
-
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
       title: body.title,
@@ -38,6 +31,20 @@ blogRouter.post('/', async (request, response) => {
   })
 
 blogRouter.delete('/:id', async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    return response
+      .status(404)
+      .json({error: 'blog does not exist'})
+  }
+
+  if ( blog.user._id.toString() !== user._id.toString()) {
+    return response.status(403).json({error: 'cannot delete blog of another user'})
+  } 
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })  
@@ -49,15 +56,29 @@ blogRouter.put('/:id', async (request, response) => {
     body.likes = 0
   }
 
-  const blog = {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    return response
+      .status(404)
+      .json({error: 'blog does not exist'})
+  }
+
+  if ( blog.user._id.toString() !== user._id.toString()) {
+    return response.status(403).json({error: 'cannot update blog of another user'}) 
+  } 
+
+  const blogUpdates = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new:true, runValidators: true})
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blogUpdates, {new:true, runValidators: true})
   response.json(updatedBlog)
+
 })
 
 module.exports = blogRouter
